+++ apps/payroll/app/routes/_protected+/employees+/$employeeId+/delete-employee.tsx
@@ -1,7 +1,14 @@
+import { DEFAULT_ROUTE } from "@/constant";
+import { safeRedirect } from "@/utils/server/http.server";
+import { getUserCookieOrFetchUser } from "@/utils/server/user.server";
 import { deleteEmployee } from "@canny_ecosystem/supabase/mutations";
 import { getSupabaseWithHeaders } from "@canny_ecosystem/supabase/server";
 import { useToast } from "@canny_ecosystem/ui/use-toast";
-import { isGoodStatus } from "@canny_ecosystem/utils";
+import {
+  deleteRole,
+  hasPermission,
+  isGoodStatus,
+} from "@canny_ecosystem/utils";
 import type { ActionFunctionArgs } from "@remix-run/node";
 import { json, useActionData, useNavigate } from "@remix-run/react";
 import { useEffect } from "react";
@@ -10,10 +17,14 @@ export async function action({
   request,
   params,
 }: ActionFunctionArgs): Promise<Response> {
+  const { supabase, headers } = getSupabaseWithHeaders({ request });
+  const { user } = await getUserCookieOrFetchUser(request, supabase);
+
+  if (!hasPermission(`${user?.role!}`, `${deleteRole}:employees`)) {
+    return safeRedirect(DEFAULT_ROUTE, { headers });
+  }
   const employeeId = params.employeeId;
   try {
-    const { supabase } = getSupabaseWithHeaders({ request });
-
     const { status, error } = await deleteEmployee({
       supabase,
       id: employeeId ?? "",
@@ -29,7 +40,7 @@ export async function action({
 
     return json(
       { status: "error", message: "Failed to delete employee", error },
-      { status: 500 },
+      { status: 500 }
     );
   } catch (error) {
     return json({
