+++ apps/payroll/app/routes/_protected+/employees+/$employeeId+/work-portfolio+/$skillId.delete-employee-skill.tsx
@@ -1,7 +1,14 @@
+import { DEFAULT_ROUTE } from "@/constant";
+import { safeRedirect } from "@/utils/server/http.server";
+import { getUserCookieOrFetchUser } from "@/utils/server/user.server";
 import { deleteEmployeeSkill } from "@canny_ecosystem/supabase/mutations";
 import { getSupabaseWithHeaders } from "@canny_ecosystem/supabase/server";
 import { useToast } from "@canny_ecosystem/ui/use-toast";
-import { isGoodStatus } from "@canny_ecosystem/utils";
+import {
+  deleteRole,
+  hasPermission,
+  isGoodStatus,
+} from "@canny_ecosystem/utils";
 import type { ActionFunctionArgs } from "@remix-run/node";
 import { json, useActionData, useNavigate } from "@remix-run/react";
 import { useEffect } from "react";
@@ -10,12 +17,17 @@ export async function action({
   request,
   params,
 }: ActionFunctionArgs): Promise<Response> {
+  const { supabase, headers } = getSupabaseWithHeaders({ request });
+  const { user } = await getUserCookieOrFetchUser(request, supabase);
+
+  if (!hasPermission(`${user?.role!}`, `${deleteRole}:employee_skills`)) {
+    return safeRedirect(DEFAULT_ROUTE, { headers });
+  }
+
   const skillId = params.skillId;
   const employeeId = params.employeeId;
 
   try {
-    const { supabase } = getSupabaseWithHeaders({ request });
-
     const { status, error } = await deleteEmployeeSkill({
       supabase,
       id: skillId ?? "",
@@ -37,7 +49,7 @@ export async function action({
         error,
         employeeId,
       },
-      { status: 500 },
+      { status: 500 }
     );
   } catch (error) {
     return json({
